+++ spotbugs/src/test/java/edu/umd/cs/findbugs/util/MutableClassesTest.java
@@ -96,6 +96,12 @@ public class MutableClassesTest {
         Assert.assertFalse(MutableClasses.mutableSignature("Ledu/umd/cs/findbugs/util/MutableClassesTest$Immutable;"));
     }
 
+    @Test
+    public void TestImmutableValuedBased() {
+        // Annotated with @jdk.internal.ValueBased and has "setValue", which should normally trip detection
+        Assert.assertFalse(MutableClasses.mutableSignature("Ljava/util/KeyValueHolder;"));
+    }
+
     @com.google.errorprone.annotations.Immutable
     public static class ErrorProneImmutable {
         public void write() {